Add files via upload
These spreedsheets give the index code for every prop that can be found in Myworld. Those marked as unknown are not accessible through the editor. Using the index code and a hex editor it is possible to locate a prop in nthe world file and replace it with another - simply by changing the code. N.B. this does not apply to props that have been painted. A more complex process must be followed in those situations.